Slight app restructure
