+++ packages/app/src/common/assets/icons/ArticlesIcon.tsx
