+++ packages/app/src/common/assets/icons/CvIcon.tsx
