+++ packages/app/src/common/assets/icons/PortfolioIcon.tsx
