+++ packages/app/src/common/assets/icons/index.ts
