+++ packages/app/src/common/components/index.ts
@@ -1,0 +1,2 @@
+export {Page} from '../../layout/Page/index.js';
+export {PageNav} from '../../layout/PageNav/index.js';
