--- packages/app/src/components/index.ts
@@ -1,2 +1,0 @@
-export {Page} from './Page';
-export {PageNav} from './PageNav';
