+++ packages/app/src/index.ts
@@ -1,3 +1,5 @@
-export {Articles} from './articles';
-export {Portfolio} from './portfolio';
-export {Resume} from './resume';
+export {Articles} from './pages/articles/index.js';
+export {Portfolio} from './pages/portfolio/index.js';
+export {Resume} from './pages/resume/index.js';
+
+export {App} from './App';
