+++ packages/app/src/layout/PageNav/PageNav.tsx
@@ -1,7 +1,7 @@
 import {Link}                                                  from '@bengsfort.dev/router';
 
-import {Github, PortfolioIcon, LinkedIn, CvIcon, ArticlesIcon} from '../../assets/icons';
-import {VisuallyHidden}                                        from '../VisuallyHidden';
+import {Github, PortfolioIcon, LinkedIn, CvIcon, ArticlesIcon} from '../../common/assets/icons/index.js';
+import {VisuallyHidden}                                        from '../../common/components/VisuallyHidden';
 
 import styles                                                  from './PageNav.module.css';
