+++ packages/router/src/index.ts
@@ -1,3 +1,3 @@
 export {Link, Route, Redirect} from './components';
-
+export type {CommonRouterProps} from './types';
 export * from './hooks';
